+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log
 ## [4.11.3](https://github.com/deltaDNA/android-sdk/releases/tag/4.11.3) 
 ### Added 
-Improved support for event triggered campaigns 
-
+- Improved support for event triggered campaigns.
+- Automatic Session Configuration retry mechanism to better deal with network related failures.
 ## [4.11.2](https://github.com/deltaDNA/android-sdk/releases/tag/4.11.2) 
 ### Fixed 
 - Evaluating non-whitelisted events for Event-Triggered Campaigns no longer fails.